fix: couldnt fetch firebase, refactor json var type

+++ lib/logic/model/firebase_data.dart
@@ -1,6 +1,6 @@
 class FirebaseData {
-  final int airHumidity;
-  final int airTemperature;
+  final double airHumidity;
+  final double airTemperature;
   final bool isRaining;
   final int lightIntensity;
   final bool roofMoverStatus;
